registration: Generate & display the QR code
Add 'generateRegistrationQR' function, that generates a QRcode with a
key, an IV and a mac address, and returns it as a string.
This function is called during enrollment, where the QR code is
displayed, in order to be scanned by a new verifier.

Also add the 'generateAES128KeyWithIV' util function, that generates
the AES key and the initial IV with the crypto/rand package.

Signed-off-by: Loic Buckwell <loic.buckwell@stagiaires.ssi.gouv.fr>