Delete existing characteristics
In order to implement control-inversion, the server will make
only one characteristic available to the clients, and choose what
to send/receive onto it on interactions.
Thus old characteristics are deleted, to prepare the introduction of
the new unique characteristic.
Currently, main.go only advertises a service, without any
characteristic.

Signed-off-by: Loic Buckwell <loic.buckwell@stagiaires.ssi.gouv.fr>