Create the ultrablue characteristic
This characteristic (ultrablueChr) abstracts the BLE transport layer.
The only interface the characteristic will expose is a go channel.
This channel is exposed through a state, that is connection related:
Each time a client interacts with the characteristic for the first time,
the state is created, and records some informations the characteristic
needs to know the read/write operation progress.

Signed-off-by: Loic Buckwell <loic.buckwell@stagiaires.ssi.gouv.fr>