Implement device registration and authentication
Implementation of the four first CBOR messages.
The two first one consist in getting the attester's
TPM endorsement key and certificate, storing them in
the room database in a device entity, and sending back
a registration confirmation.
The two others are for authentication. We get a challenge
nonce, and send it back so that the server can authenticate
the verifier device.

Signed-off-by: Loic Buckwell <loic.buckwell@stagiaires.ssi.gouv.fr>