Fix issue #10, remove the registration part
- Fix the issue that makes the server crash if a
virtual TPM was used. Now we check that the EK
certificate exists before dereferencing it.

- The registration was made on successful EK exchange.
If an error arised after that, invalid devices would have
been stored on both sides.
This part has gone, and now we register both devices on both
side at the end of the protocol: in the 'response' protocol
function.

Signed-off-by: Loic Buckwell <loic.buckwell@stagiaires.ssi.gouv.fr>