Rename ekenroll function to sendRegistrationData
Now, the client decides wether or not to send back a secret
to extend the 9th PCR. This implies that the server must
provide him a way to know if he must send one for later
attestations during enrollment. This is done through a
boolean value passed alongside the EK pub/cert in the first
enrollment message.
Thus, this message no longer contains only an EK, but also
other "registration data". To be consistent, this function
has been renamed.

Signed-off-by: Loic Buckwell <loic.buckwell@stagiaires.ssi.gouv.fr>