[android] Implement encryption/decryption
At enroll time, when the QRcode is scanned, we now
get the symmetric key inside, which is passed to the
protocol view.

Two more steps are added to the protocol:
- UUID_SEND: The first message to send to the server is now
  the client UUID, which is generated at enroll time, or retrieved
  from the database device object for an attestation
- ENABLE_ENCRYPTION: Once we sent the client UUID, we can start
  encrypting communications by calling the enableEncryption callback.
  This will create a CryptoManager instance, with the key obtained from
  the qr code at enroll time or from the keystore for an attestation.

The hooks for sending/receiving messages are thus updated so that
if a new cryptoManager variable is non-null, messages are encrypted before
being sent, and decrypted on reception.

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>