[server] Add encryption key persistence
At enroll time, the key is saved in the form of a TPM sealed
object in two files:

- /etc/ultrablue/CLIENT_UUID
- /etc/ultrablue/CLIENT_UUID.pub

It is sealed to the Storage Root Key (SRK), so that if either the
TPM or the TPM owner is modified, those files couldn't be decrypted
anymore.

When running an attestation, the blobs are read from those files and
the resulting object is unsealed to retrieve the key.

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>