[server] Allow to provide a PIN to seal keys
Add a --with-pin option to the ultrablue server.
When used:

- At enroll time: The user is prompted to choose a PIN
to seal the encryption key on disk

- For an attestation: The PIN must be given
back to retrieve the encryption key

If the option is omitted for an attestation whereas it was
provided at enroll time, the policy will fail but no unsealing
attempt is made, thus the TPM fail counter is not incremented.

If the option is given for an attestation whereas it wasn't at
enroll time, the user will be prompted to enter a PIN, but any
PIN will do the job as it will not be used to unseal the key.

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>