[server] Authentication nonce tweak
Fix a potential authentication bypass:

During the authentication phase, we send an encrypted nonce,
with a prepended IV, needed by the client to decrypt the nonce
with the shared key. The client then have to reencrypt the nonce
with another IV and send it back to the server.
The server then compares the received nonce with the sent one
and if they match, the client is authenticated.

The issue here is that an attacker can connect, listen for
the encrypted nonce from the server, and send back the exact same
message. This message contains the right nonce, encrypted with the
shared key and a valid IV.

Tweaking the nonce in a deterministic way before sending it back ensures
that someone who doesn't know the shared key can't successfully bypass the
authentication process.

The way we tweak the nonce is the following:

- We split the nonce in two halfs (8 bytes long each)
- We invert those halfs

In python, it would give: tweaked_nonce = nonce[8:] + nonce[:8]

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>