[Android] Send an attestation error on any post quote error
This isn't clear yet, but we have to choose from which point we want to
make the attestation fail if an error occurs.

Before that point, any error will simply close the connection, and the
server will expect another client to connect, to restart the
attestation.
After that point, if an error occurs, the server will assume the
attestation failed, and will exit. If ultrablue is configured to do
disk decryption at boot time, it means that a fallback LUKS keyslot will be
used (if there is any) to decrypt the disk.

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>