[android] Prevent crash on device renaming
When navigating to a device fragment, going back
to main menu and navigating again to a device fragment,
hitting the edit button resulted in a crash, because
the context is null.

This commit avoids to dereference the null context and prevents
the application to crash, but does not fix the bug.
I still need to figure out why the context is trashed on second
and later fragment creation.

Signed-off-by: Loic Buckwell <lfalkau@student.42.fr>