[ios] Implement communication encryption
This commit implements the protocol communication encryption and decryption.
Protocol implications are described in the documentation, and in the server commits.

On the IOS client, encryption keys are stored in the keystore. Note that this is a work
in progress feature, and some things are not doing well for now. (e.g. the encryption keys
aren't deleted on device deletion).

The commit would need some rework, I did it this way because of time constraints. It aims
at support the SSTIC submission, providing a working ios client POC for the current server.

Signed-off-by: Loïc Buckwell <lfalkau@student.42.fr>